vm: widen function values to 48-bit payload; 27-bit closure id, 20-bit chunk idx
The 32-bit packing (closure_id << 16 | chunk_idx, as i32) overflowed
into the native-sentinel space at 32 768 closures: the sign bit set,
every closure created after that dispatched as a garbage builtin and
returned undefined. A single heavy page crosses that line mid-load —
DuckDuckGo's SERP creates ~100 accessor closures per DOM element
wrapper, which is how React's createRoot(container) saw a container
whose .nodeType getter 'returned' undefined and died with error #299.
Chunk indices were also two-thirds of the way to their own 16-bit
cliff (17 085 chunks on that page).

Function payloads are now sign-extended 48-bit: negative = sentinel,
non-negative = closure_id << 20 | chunk_idx via Value::pack_closure /
fn_closure_id / fn_chunk_idx. Sentinel plumbing widened i32 → i64.

test262: 26763/29080 (unchanged). decode bench 0.70s (baseline 0.75s).
All copper smokes green; DDG SERP mounts React past createRoot.

+++ src/runtime/object.rs
@@ -277,7 +277,7 @@ pub enum FunctionKind {
         args: Vec<Value>,
     },
     /// Native sentinel function (negative i32 ID), for binding native methods
-    NativeSentinel { sentinel: i32 },
+    NativeSentinel { sentinel: i64 },
 }
 
 /// Extract an ObjectId from a Value, if it references one.
